Make catkey an attribute on SolrDocument
